requirements.txt
pymongo==3.5.0
numpy==1.16.1
progressbar2==3.34.2
matplotlib==2.2.4
scipy==0.19.1
requests==2.18.3